app.py del

+++ 0x03-user_authentication_service/app.py
@@ -58,16 +58,17 @@ def logout() -> str:
     Return:
         - Redirects to home route.
     """
-    session_id = request.cookies.get("session_id", None)
-    if session_id is None:
-        abort(403)
-    try:
-        user = AUTH.get_user_from_session_id(session_id)
-        if user:
-            AUTH.destroy_session(user.id)
-            return redirect("/")
-    except Exception:
-        abort(403)
+    if request.method == "DELETE":
+        session_id = request.cookies.get("session_id", None)
+        if session_id is None:
+            abort(403)
+        try:
+            user = AUTH.get_user_from_session_id(session_id)
+            if user:
+                AUTH.destroy_session(user.id)
+                return redirect("/")
+        except Exception:
+            abort(403)
 
 
 @app.route('/profile', methods=['GET'], strict_slashes=False)
